Refactoring ,Abstraction and remove unnecessary code Relates #32
