+++ public/css/style.css
@@ -75,10 +75,6 @@ body {
   box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
 }
 
-.meal-name {
-  /* font-size: 2.3vw; */
-}
-
 .meal-img-div {
   width: 100%;
 }
@@ -195,9 +191,6 @@ body {
 .contant_box_404 p {
   color: #000000
 }
-
-
-
 
 
 /*======================
