+++ public/js/Dom.js
@@ -1,0 +1,3 @@
+document.querySelector('form').addEventListener('submit', (e) => {
+  // e.preventDefault();
+});
