--- public/js/app.js
@@ -1,19 +1,0 @@
-
-// document.querySelector('.search-btn').addEventListener('click', () => {
-// document.querySelector('.result').textContent = '';
-// const query = JSON.stringify({ input: document.querySelector('.input').value });
-//   fetch('/search', {
-//     method: 'POST',
-//     body: query,
-//     headers: { 'Content-Type': 'application/json' },
-//   }).then(res => res.text()).then((res) => {
-//     document.open();
-//     document.write(res);
-//     document.close();
-//   })
-//     .catch(error => console.log(error));
-// });
-document.querySelector('form').addEventListener('submit', (e) => {
-
-  // e.preventDefault();
-});
