+++ src/app.js
@@ -1,22 +1,15 @@
 const express = require('express');
-
 const exphbs = require('express-handlebars');
-
 const path = require('path');
 const { Ingredient } = require('./views/helpers/ingredient');
-
-
-const routes = require('./routes');
+const routes = require('./Controllers');
 
 const app = express();
 
 app.use(express.static(path.join(__dirname, '..', 'public')));
-
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
-// helper
 
-// // you need to access with hbs files
 app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'hbs');
 app.engine(
