+++ src/views/404.hbs
@@ -1,25 +1,22 @@
     <section class="page_404">
-        <div class="container">
-          <div class="row">	
+      <div class="container">
+        <div class="row">
           <div class="col-sm-12 ">
-          <div class="col-sm-10 col-sm-offset-1  text-center">
-          <div class="four_zero_four_bg">
-            <h1 class="text-center ">404</h1>
-          
-          
-          </div>
-          
-          <div class="contant_box_404">
-          <h3 class="h2">
-          Look like you're lost
-          </h3>
-          
-          <p>the page you are looking for not avaible!</p>
-          
-          <a href="/" class="link_404">Back to Home</a>
-        </div>
-          </div>
-          </div>
+            <div class="col-sm-10 col-sm-offset-1  text-center">
+              <div class="four_zero_four_bg">
+                <h1 class="text-center ">404</h1>
+              </div>
+              <div class="contant_box_404">
+                <h3 class="h2">
+                  Look like you're lost
+                </h3>
+
+                <p>the page you are looking for not avaible!</p>
+
+                <a href="/" class="link_404">Back to Home</a>
+              </div>
+            </div>
           </div>
         </div>
-      </section>
+      </div>
+    </section>
