+++ src/views/home.hbs
@@ -26,9 +26,7 @@
         <a class="youtubeLink" href="{{this.strYoutube}}">Youtube</a>
         <a class="youtubeLink" href={{this.strSource}}>Source Link</a>
       </div>
-
     </div>
-       
       {{/each}}
   </div>
 </main>
